Reduce isolated folders in model garden. Move movielens to recommendation
PiperOrigin-RevId: 269680664
