+++ official/recommendation/data_pipeline.py
@@ -35,8 +35,8 @@ from six.moves import queue
 import tensorflow as tf
 from absl import logging
 
-from official.datasets import movielens
 from official.recommendation import constants as rconst
+from official.recommendation import movielens
 from official.recommendation import popen_helper
 from official.recommendation import stat_utils
