+++ official/recommendation/data_test.py
@@ -27,9 +27,9 @@ import numpy as np
 import scipy.stats
 import tensorflow as tf
 
-from official.datasets import movielens
 from official.recommendation import constants as rconst
 from official.recommendation import data_preprocessing
+from official.recommendation import movielens
 from official.recommendation import popen_helper
 from official.utils.misc import keras_utils
